requirements.txt
keyboard
pyautogui
python-dotenv